Throw exact error message when get files
When object request fails, the current exception message is
"The specified file doesn't exist." always.
After change, the error message of the AWS outcome will be
thrown with a runtime_error in C++. Note that the user cannot
catch this error in Python because there's not an interface
between Python and C++ for exceptions.