change S3Dataset to mapped-style
Before S3Dataset class inherits PyTorch IterableDataset, and when iterate
the dataset, samples will be returned. However, the returned samples are not in
consistent format. Specifically:
1. if it’s a single file (csv, jpeg etc.) the file is directly returned as
Byteio format. (not a filename fileobj pair)
2. if it’s a tar/zip, a pair of (filename, fileobj) will be returned.
Specifically, the filename is the name inside the tar/zip file, and the fileobj
is returned after io.ByteIO(tar/zip file) and the file with read()

After changing to a mapped-style dataset, S3dataset is:
1. Added `__getitem__()` function, so `dataset[idx]` will return a pair of
(filename, fileobj) with `idx`th index in the `self.urls_list`. Note that this
is not the `idx`th filename in the input `urls_list`, because there can be
prefixes in the input `urls_list`.
2. the return `fileobj` will be in the raw byteio format, because that's an
acceptable format regardless of tar, csv, jpeg, zip, or any other format.
3. tests are also changed based on this new consistent output api.