Removes the check for num_batches
batch count was maintained in the for iteration over the dataloader. In case of S3IterableDataset, that loop is executed bu number of worker times
Hence the variable num_batches did not hold the actual number of batches.
This is removed as loading all data correctly is suffient to establish correct working of S3 plugin