Fix hanging issue with num_workers > 0
When num_workers in data-loader is greater than zero, the workers are hung
during AwsInitOptions() call. It turns out we had enabled log level of tracing
which leads to main process holding locks on the trace file. When workers call
AwsInitOptions() they get stuck because they also try to acquire the lock on
the same trace file. The fix is to disable trace level logging which is anyway
not required and have each worker has it's own handler to make s3 calls.

Test Coverage
All tests are passing including integration tests.