Add an automated integration test
These changes will allow the user to launch an EC2 instance
and run the CV integration test that we had used during the
implementation of the project. The script will send commands
to the instance and terminate it once the test is complete.
The results are saved in an S3 directory.

It should be noted that the integ_test_cfg.py that stores
information needed to our specific test has been removed.

This commit includes a README with instructions for running
the test.